requirements.txt
ultralytics
opencv-python
numpy
scipy
pyyaml
confluent-kafka